Split `i2c` module into sub-modules
This follows the example set in `usart` and other modules. It's not
strictly necessary (`i2c` is a relatively small module), but I think
it's still an improvement.
